Fixed rendering issue in Gateware's GRasterSurface
GRasterSurface in Gateware.h was preventing the display buffer from being rendered with SmartUpdateSurface(). This is because of two separate checks the function was making to ensure the display buffer size didn't exceed or was equal to the size of the window client area. However, if the display buffer size is equal to the size of the client area then it should proceed with rendering. The conditional checks have been updated so they ensure the display buffer doesn't exceed the size of the window's client area only.

This issue has been added to Gateware's issue tracker: gitlab.com/gateware-development/gateware/-/issues/139